Streamline error handling / view.
Pull all error handling into the same view.  There are a lot of
different types of erros and I don't want to subclass a view for each
one as was the current design.  Pulling this into ErrorView.

Fixing various exception handling routines.

Chaged Message(model).getDescription to getMeta;  Returns array of meta
content abou the message (errors included).  This is used to parse out
a messages display and handling in the view in a consistent way as well
as to give the nav bar a single way to show the "last message".  Before
this was being done in several different ways.