Optimized prekey handling.
There were some issues in how stale device detection was intermixing
with normal prekey lookups for new devices.  The /* optimization was
never being used as a result.

We used to have to fetch prekeys one at a time because it lessened the
likelyhood of the signal server from blowing up due to a postgres
db issue.  I fixed the signal server to handle these, so it's safe
to fetch them as fast as the network allows now.